Extract stackedAreaChart fixtures from testData
